test(e2e): extract shared web e2e REST helpers to support/api
