+++ frontend/e2e/note-live-update.spec.ts
@@ -1,4 +1,5 @@
-import { expect, type Page, test } from "@playwright/test";
+import { expect, test } from "@playwright/test";
+import { createVault, registerAndLogin, signInForNote, upsertNote } from "./support/api";
 
 /**
  * #277 — Two-session live-update behavior for the web SPA note editor.
@@ -13,123 +14,6 @@ import { expect, type Page, test } from "@playwright/test";
  * note_changed event that re-seeds the Y.Doc; true concurrent in-browser edits
  * converge via the CRDT protocol.
  */
-
-const PASS = "E2eTestPass!99";
-
-async function registerAndLogin(baseURL: string, email: string): Promise<string> {
-	const reg = await fetch(`${baseURL}/api/auth/register`, {
-		method: "POST",
-		headers: { "Content-Type": "application/json" },
-		body: JSON.stringify({ email, password: PASS }),
-	});
-	if (!reg.ok && reg.status !== 422) {
-		throw new Error(`register failed: ${reg.status} ${await reg.text()}`);
-	}
-
-	const login = await fetch(`${baseURL}/api/auth/login`, {
-		method: "POST",
-		headers: { "Content-Type": "application/json" },
-		body: JSON.stringify({ email, password: PASS }),
-	});
-	if (!login.ok) {
-		throw new Error(`login failed: ${login.status} ${await login.text()}`);
-	}
-	const { access_token } = (await login.json()) as { access_token: string };
-
-	const auth = { "Content-Type": "application/json", Authorization: `Bearer ${access_token}` };
-	const prof = await fetch(`${baseURL}/api/onboarding/profile`, {
-		method: "PATCH",
-		headers: auth,
-		body: JSON.stringify({ uses_obsidian: true, tools: ["claude"] }),
-	});
-	if (!prof.ok) {
-		throw new Error(`onboarding PATCH failed: ${prof.status} ${await prof.text()}`);
-	}
-	// Vault is created later by the spec (createVault helper). Just suppress
-	// the checklist tour row so the dashboard doesn't intercept editor clicks.
-	const act = await fetch(`${baseURL}/api/onboarding/actions`, {
-		method: "POST",
-		headers: auth,
-		body: JSON.stringify({ action: "dismissed:tour" }),
-	});
-	if (!act.ok) {
-		throw new Error(`onboarding action POST failed: ${act.status} ${await act.text()}`);
-	}
-
-	return access_token;
-}
-
-interface Vault {
-	id: number;
-	name: string;
-}
-
-async function createVault(baseURL: string, token: string, name: string): Promise<Vault> {
-	const res = await fetch(`${baseURL}/api/vaults`, {
-		method: "POST",
-		headers: { "Content-Type": "application/json", Authorization: `Bearer ${token}` },
-		body: JSON.stringify({ name }),
-	});
-	if (!res.ok) {
-		throw new Error(`vault create failed: ${res.status} ${await res.text()}`);
-	}
-	const { vault } = (await res.json()) as { vault: Vault };
-	return vault;
-}
-
-async function upsertNote(
-	baseURL: string,
-	token: string,
-	vaultId: number,
-	path: string,
-	content: string,
-	version?: number,
-): Promise<{ id: number }> {
-	const res = await fetch(`${baseURL}/api/notes`, {
-		method: "POST",
-		headers: {
-			"Content-Type": "application/json",
-			Authorization: `Bearer ${token}`,
-			"X-Vault-Id": String(vaultId),
-		},
-		body: JSON.stringify({ path, content, mtime: Date.now() / 1000, version }),
-	});
-	if (!res.ok) {
-		throw new Error(`note upsert failed: ${res.status} ${await res.text()}`);
-	}
-	const { note } = (await res.json()) as { note: { id: number } };
-	return { id: note.id };
-}
-
-/**
- * Sign in by going to the target note URL first — AuthGuard redirects to
- * `/sign-in?return_to=...` and bounces back after sign-in completes.
- * Seed `engram.activeVaultId` on the sign-in page so it survives the post-
- * sign-in route change (same origin, same storage).
- */
-async function signInForNote(
-	page: Page,
-	email: string,
-	vaultId: number,
-	noteId: number,
-): Promise<void> {
-	// Hitting /note/:id unauth -> AuthGuard redirects to /sign-in?return_to=...
-	await page.goto(`/note/${noteId}`);
-	await expect(page).toHaveURL(/\/sign-in/u, { timeout: 10_000 });
-
-	// Seed active vault before sign-in completes so the post-redirect render
-	// already has it (vault-switcher's auto-select wouldn't pick our new vault
-	// before NotePage's first query fires).
-	await page.evaluate((id) => {
-		localStorage.setItem("engram.activeVaultId", String(id));
-	}, vaultId);
-
-	await page.getByLabel("Email").fill(email);
-	await page.getByLabel("Password", { exact: true }).fill(PASS);
-	await page.getByRole("button", { name: /sign in/iu }).click();
-
-	await expect(page).toHaveURL(new RegExp(`/note/${noteId}`), { timeout: 10_000 });
-}
 
 test.describe("SPA viewer live-update (#277)", () => {
 	test("viewer re-renders when a remote client upserts the open note", async ({
